+++ spec/jwt_spec.rb
@@ -38,7 +38,7 @@ RSpec.describe JWT do
       'PS512' => ''
     }
 
-    if ::JWT.rbnacl?
+    if JWT.rbnacl?
       ed25519_private = RbNaCl::Signatures::Ed25519::SigningKey.new('abcdefghijklmnopqrstuvwxyzABCDEF')
       ed25519_public =  ed25519_private.verify_key
       data.merge!(
@@ -267,7 +267,7 @@ RSpec.describe JWT do
     end
   end
 
-  if ::Gem::Version.new(OpenSSL::VERSION) >= ::Gem::Version.new('2.1')
+  if Gem::Version.new(OpenSSL::VERSION) >= Gem::Version.new('2.1')
     %w[PS256 PS384 PS512].each do |alg|
       context "alg: #{alg}" do
         before(:each) do
@@ -621,7 +621,7 @@ RSpec.describe JWT do
   context 'when the alg value is given as a header parameter' do
     it 'does not override the actual algorithm used' do
       pending 'Breaking change in 3.0'
-      headers = ::JSON.parse(Base64.urlsafe_decode64(JWT.encode('Hello World', 'secret', 'HS256', { alg: 'HS123' }).split('.').first))
+      headers = JSON.parse(JWT::Base64.url_decode(JWT.encode('Hello World', 'secret', 'HS256', { alg: 'HS123' }).split('.').first))
       expect(headers['alg']).to eq('HS256')
     end
 
@@ -632,7 +632,7 @@ RSpec.describe JWT do
 
   context 'when hmac algorithm is used without secret key' do
     it 'encodes payload' do
-      pending 'Different behaviour on OpenSSL 3.0 (https://github.com/openssl/openssl/issues/13089)' if ::JWT.openssl_3_hmac_empty_key_regression?
+      pending 'Different behaviour on OpenSSL 3.0 (https://github.com/openssl/openssl/issues/13089)' if JWT.openssl_3_hmac_empty_key_regression?
       payload = { a: 1, b: 'b' }
 
       token = JWT.encode(payload, '', 'HS256')
@@ -667,9 +667,9 @@ RSpec.describe JWT do
 
   describe '::JWT.decode with verify_iat parameter' do
     let!(:time_now) { Time.now }
-    let(:token)     { ::JWT.encode({ pay: 'load', iat: iat }, 'secret', 'HS256') }
+    let(:token)     { JWT.encode({ pay: 'load', iat: iat }, 'secret', 'HS256') }
 
-    subject(:decoded_token) { ::JWT.decode(token, 'secret', true, verify_iat: true) }
+    subject(:decoded_token) { JWT.decode(token, 'secret', true, verify_iat: true) }
 
     before { allow(Time).to receive(:now) { time_now } }
 
@@ -690,7 +690,7 @@ RSpec.describe JWT do
     context 'when iat is 1 second before Time.now' do
       let(:iat) { time_now.to_i + 1 }
       it 'raises an error' do
-        expect { decoded_token }.to raise_error(::JWT::InvalidIatError, 'Invalid iat')
+        expect { decoded_token }.to raise_error(JWT::InvalidIatError, 'Invalid iat')
       end
     end
   end
@@ -792,7 +792,7 @@ RSpec.describe JWT do
     let(:key_finder) { instance_double('::JWT::X5cKeyFinder') }
 
     before do
-      expect(::JWT::X5cKeyFinder).to receive(:new).with(root_certificates, nil).and_return(key_finder)
+      expect(JWT::X5cKeyFinder).to receive(:new).with(root_certificates, nil).and_return(key_finder)
       expect(key_finder).to receive(:from).and_return(data[:rsa_public])
     end
 
@@ -832,23 +832,38 @@ RSpec.describe JWT do
   end
 
   describe 'when none token is and decoding without key and with verification' do
-    let(:none_token) { ::JWT.encode(payload, nil, 'none') }
+    let(:none_token) { JWT.encode(payload, nil, 'none') }
     it 'decodes the token' do
-      expect(::JWT.decode(none_token, nil, true, algorithms: 'none')).to eq([payload, { 'alg' => 'none' }])
+      expect(JWT.decode(none_token, nil, true, algorithms: 'none')).to eq([payload, { 'alg' => 'none' }])
     end
   end
 
   describe 'when none token is decoded with a key given' do
-    let(:none_token) { ::JWT.encode(payload, nil, 'none') }
+    let(:none_token) { JWT.encode(payload, nil, 'none') }
     it 'decodes the token' do
-      expect(::JWT.decode(none_token, 'key', true, algorithms: 'none')).to eq([payload, { 'alg' => 'none' }])
+      expect(JWT.decode(none_token, 'key', true, algorithms: 'none')).to eq([payload, { 'alg' => 'none' }])
     end
   end
 
   describe 'when none token is decoded without verify' do
-    let(:none_token) { ::JWT.encode(payload, nil, 'none') }
+    let(:none_token) { JWT.encode(payload, nil, 'none') }
     it 'decodes the token' do
-      expect(::JWT.decode(none_token, 'key', false)).to eq([payload, { 'alg' => 'none' }])
+      expect(JWT.decode(none_token, 'key', false)).to eq([payload, { 'alg' => 'none' }])
+    end
+  end
+
+  describe 'when token signed with nil and decoded with nil' do
+    let(:no_key_token) { JWT.encode(payload, nil, 'HS512') }
+    it 'raises JWT::DecodeError' do
+      pending 'Different behaviour on OpenSSL 3.0 (https://github.com/openssl/openssl/issues/13089)' if JWT.openssl_3_hmac_empty_key_regression?
+      expect { JWT.decode(no_key_token, nil, true, algorithms: 'HS512') }.to raise_error(JWT::DecodeError, 'No verification key available')
+    end
+  end
+
+  context 'when token ends with a newline char' do
+    let(:token) { "#{JWT.encode(payload, 'secret', 'HS256')}\n" }
+    it 'ignores the newline and decodes the token' do
+      expect(JWT.decode(token, 'secret', true, algorithm: 'HS256')).to include(payload)
     end
   end