+++ CHANGELOG.md
@@ -17,6 +17,20 @@
 - Algorithms moved under the `::JWT::JWA` module ([@anakinj](https://github.com/anakinj))
 - Your contribution here
 
+## [v2.7.2](https://github.com/jwt/ruby-jwt/tree/v2.7.2) (NEXT)
+
+[Full Changelog](https://github.com/jwt/ruby-jwt/compare/v2.7.1...v2.7.2)
+
+**Features:**
+
+- Updated rubocop to 1.56 [#573](https://github.com/jwt/ruby-jwt/pull/573) - [@anakinj](https://github.com/anakinj).
+- Your contribution here
+
+**Fixes and enhancements:**
+
+- Fix signature has expired error if payload is a string [#555](https://github.com/jwt/ruby-jwt/pull/555) - [@GobinathAL](https://github.com/GobinathAL).
+- Your contribution here
+
 ## [v2.7.1](https://github.com/jwt/ruby-jwt/tree/v2.8.0) (2023-06-09)
 
 [Full Changelog](https://github.com/jwt/ruby-jwt/compare/v2.7.0...v2.8.0)